Add files via upload
Here are all the datasheets you might need to understand how the LED's operate and the hardware that is being used to control them. The PDF's with MPLAB in the title are for the program that is being used to program the microcontroller, the PIC datasheet is for the microcontroller, and the other two are for understanding how the LED strips operate.